Admin dash functionality

+++ flaskr/auth.py
@@ -5,15 +5,15 @@ since been rewritten constantly to suit our needs. at this point this is
 """
 
 # auth.py
-
-from flask import render_template, redirect, flash, url_for
 from flask_login import login_user, logout_user, current_user
-from werkzeug.security import generate_password_hash, check_password_hash
-from .models import db, Users
 from .file_upload import *
 from .forms import LoginForm, RegistrationForm
 from .search import searchBar
-from .accountManagement import uploadUser
+from .accountManagement import *
+import csv
+import os
+from werkzeug.utils import secure_filename
+from flask import current_app
 
 # Home route
 def home():
@@ -21,7 +21,7 @@ def home():
         return redirect(url_for('landing'))
     form1 = get_showform(1)
     form2 = get_showform(2)
-    return render_template("file-upload/file_upload.html", show_form1=form1, show_form2=form2)
+    return render_template("file-upload/file_upload.html", show_form1=form1, show_form2=form2, isAdmin=current_user.administrator)
 
 # Landing route
 from flask import render_template, redirect, url_for, flash
@@ -102,22 +102,10 @@ def faculty():
 def dashboard():
     user = Users.query.filter_by(username=current_user.username).first()
     if user.administrator:
-        userlist = searchBar(request.form.get('user_search', ""), "username")
-        if request.method == 'POST':    # Detects if search bar is used at all
-            if request.form["formType"] == "user_data":
-                print("ack")
-            elif request.form["formType"] == "email":   #Search through email bar
-                emaillist = searchBar(request.form.get('email_search', ""), "email")
-                return render_template('user-tools/strappedDash.html', userlist=emaillist, warning=0)
-            elif request.form["formType"] == "username":  #search through username bar
-                userlist = searchBar(request.form.get('user_search', ""), "username")
-                return render_template('user-tools/strappedDash.html', userlist=userlist, warning=0)
-            else:   #reset search
-                userlist = searchBar("", "reset")
-                return render_template('user-tools/strappedDash.html', userlist=userlist, warning=0)
+        userlist = searchBar("", "reset")
         return render_template('user-tools/strappedDash.html', userlist=userlist, warning=0)
     else:
-        return render_template("file-upload/file_upload.html", show_form1=show_form1, show_form2=show_form2)
+        return render_template("file-upload/file_upload.html", show_form1=show_form1, show_form2=show_form2, isAdmin=False)
 
 def searchuser():
     if request.form["formType"] == "email":   #Search through email bar
@@ -138,7 +126,7 @@ def adduser():
         return render_template('user-tools/strappedDash.html', userlist=userlist, warning=1)
     else:
         uploadedUserData = [request.form.get('unEnter', ""), request.form.get('emEnter', ""), request.form.get('pwEnter', ""), request.form.get('admEnter', ""), request.form.get('admEnter', "")]
-        print(request.form.get('admEnter', ""))
+        #print(request.form.get('admEnter', ""))
         uploadUser(uploadedUserData)
         userlist = searchBar("", "reset")
         return render_template('user-tools/strappedDash.html', userlist=userlist, warning=2)
@@ -150,8 +138,57 @@ def deleteuser():
     #As of now, foundUser is a list with a list with the id
     for a in foundUser:
         for b in a:
-            Users.query.filter_by(id=b).delete()
-            db.session.commit()
+            #refuses to delete account if the account being deleted is the one using to delete
+            #basically, you cannot delete yourself
+            if b == current_user.id:
+                userlist = searchBar(request.form.get('user_search', ""), "username")
+                return render_template('user-tools/strappedDash.html', userlist=userlist, warning=4)
+            #otherwise, continue through with deletion
+            else:
+                Users.query.filter_by(id=b).delete()
+                db.session.commit()
 
     userlist = searchBar(request.form.get('user_search', ""), "username")
-    return render_template('user-tools/strappedDash.html', userlist=userlist, warning=False)
+    return render_template('user-tools/strappedDash.html', userlist=userlist, warning=3)
+
+def csvupload():
+    rejectFlag = 2
+    rejectCount = 0
+    userCount = 0
+    #thanks matt
+    #THIS, is a CRAFTING TABLE. (save the file)
+    if request.method == 'POST':
+        file_link = request.files['filename']
+        file_link.save(file_link.filename)
+        name =  file_link.filename
+
+    #first, we MINE! (let's verify the file itself before we parse it)
+        if not verifyCSV(name): #if not the correct file,
+            os.remove(name) #remove incorrect file
+            userlist = searchBar("", "reset")
+            return render_template('user-tools/strappedDash.html', userlist=userlist, warning=1)
+
+        os.rename(name, "flaskr/static/userupload.csv")
+
+    #then we CRAFT! (parse the csv file data)
+        with open("flaskr/static/userupload.csv", newline='') as csvfile:
+            listOfUsers = csv.reader(csvfile, delimiter=',', quotechar='|')
+
+    #flint and STEEL! (we check to see if the data itself is valid, specifically the username)
+            for row in listOfUsers:
+                userCount += 1
+                if Users.query.filter_by(username=row[0]).first():  #reject if list has an existing user already
+                    print("no it is now ILLEGAL")
+                    rejectFlag = 5
+                    rejectCount += 1
+                else:
+                    uploadUser(row)
+
+    #the NETHER! (get rid of the file since we are done parsing it)
+    os.remove("flaskr/static/userupload.csv")
+    userlist = searchBar("", "reset")
+    return render_template('user-tools/strappedDash.html', userlist=userlist, warning=rejectFlag, rejectedNumber=rejectCount, totalNumber=userCount)
+    #i havent even watched the minecraft movie but these sound bits are rotting my brain as we speak
+
+# def edituser():
+# will be used later for when editing an already existing user
